// src/main/resources/APachas/src/app/models/MProduct.ts
export class MProduct {
    productId: number;
    productName: string;
    productDescription: string;
    productCost: number;
    productQuantity: number;
    productPhoto: string;
    eventId: number;
    productPurchaseDate: string;
    productActive: boolean;
    productCreation: Date;
    productRemoval: Date;

}